+++ src/main/java/com/kaya/kaya01/repository/CategorieRepository.java
@@ -3,5 +3,5 @@ package com.kaya.kaya01.repository;
 import com.kaya.kaya01.Entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategorieRepository extends JpaRepository<Category, Integer> {
+public interface CategorieRepository extends JpaRepository<Category, Long> {
 }
