+++ src/main/java/com/kaya/kaya01/repository/PropertyRepository.java
@@ -3,5 +3,5 @@ package com.kaya.kaya01.repository;
 import com.kaya.kaya01.Entity.Property;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PropertyRepository extends JpaRepository<Property, Integer> {
+public interface PropertyRepository extends JpaRepository<Property, Long> {
 }
