+++ src/main/java/com/kaya/kaya01/repository/SearchRepository.java
@@ -3,5 +3,5 @@ package com.kaya.kaya01.repository;
 import com.kaya.kaya01.Entity.Search;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SearchRepository extends JpaRepository<Search, Integer> {
+public interface SearchRepository extends JpaRepository<Search, Long> {
 }
