+++ src/main/java/com/kaya/kaya01/service/UserService.java
@@ -7,10 +7,10 @@ import java.util.List;
 public interface UserService {
 
     UserDTO createUser(UserDTO userDTO);
-    UserDTO updateUserById(Integer id,UserDTO userDTO);
-    UserDTO findUserById(Integer id);
+    UserDTO updateUserById(Long id,UserDTO userDTO);
+    UserDTO findUserById(Long id);
     List<UserDTO> findAllUser();
-    void  deleteUser(Integer id);
+    void  deleteUser(Long id);
 
 
 }
